+++ src/commonMain/kotlin/csense/kotlin/extensions/collections/list/List.kt
@@ -3,6 +3,7 @@
 package csense.kotlin.extensions.collections.list
 
 import csense.kotlin.annotations.numbers.IntLimit
+import csense.kotlin.extensions.InvokeIsInstance
 import csense.kotlin.extensions.primitives.isNegativeOrZero
 
 /**
@@ -92,4 +93,41 @@ inline fun <reified T> List<T>.repeatToSize(@IntLimit(from = 0) toSize: Int): Li
     val resultList = this.repeat(timesToRepeat)
     //and add the sublist missing part.
     return resultList + this.subList(0, missingItemsToCopy)
+}
+
+/**
+ * invokes the given action on each item that is of the expected type (U)
+ * @receiver List<*>
+ * @param indices IntProgression the indexes to go over.
+ * @param action Function1<U, *>
+ */
+inline fun <reified U> List<*>.forEachIsInstance(
+        indices: IntProgression,
+        action: Function1<U, *>
+) = indices.forEach {
+    getOrNull(it)?.InvokeIsInstance(action)
+}
+
+/**
+ * invokes the given action on each item that is of the expected type (U)
+ * @receiver List<*>
+ * @param action Function1<U, *> action to invoke if the element is of type U
+ */
+inline fun <reified U> List<*>.forEachIsInstance(
+        action: Function1<U, *>
+) = forEach {
+    it.InvokeIsInstance(action)
+}
+
+
+/**
+ * Combines the inner list of the other with this.
+ * @receiver List<List<T>>
+ * @param other List<List<T>>
+ * @return List<List<T>>
+ */
+fun <T> List<List<T>>.combine(
+        other: List<List<T>>
+): List<List<T>> = mapIndexed { index, list ->
+    list + other.getOrNull(index).orEmpty()
 }